requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
dj-database-url==0.4.2
Django==2.0.2
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
pkg-resources==0.0.0
psycopg2==2.7.4
python-dotenv==0.7.1
pytz==2018.3
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
